Fix clippy warnings for wasm32-wasip1 target

+++ src/uucore/src/lib/features/fs.rs
@@ -188,7 +188,7 @@ impl FileInformation {
         #[cfg(windows)]
         return self.0.number_of_links();
         #[cfg(target_os = "wasi")]
-        return self.0.st_nlink as u64;
+        return self.0.st_nlink;
     }
 
     #[cfg(unix)]
